backend/requirements.txt
# Python 3.11+ / 3.13 compatible (raw SQL via psycopg, no SQLAlchemy)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psycopg[binary]>=3.2
sqlalchemy>=2.0.0
pydantic[email]>=2.10.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
greenlet>=3.0.0
